move utils classes to subdirectory
