+++ Profiler/Log.cpp
@@ -4,7 +4,7 @@
 #include <fstream>
 #include <algorithm>
 #include <winuser.h>
-#include "WindowsUtils.h"
+#include "utils/WindowsUtils.h"
 
 static const int BUFFER_SIZE = 2048;
