+++ Profiler/utils/WindowsUtils.h
